CSV to QIF
Converts one specific investment CSV file output from a specific 401(k) retirement
account, to Quicken Interchange Format for import into Quicken.  The specificity
implies that it's not nearly robust enough to handle any given export/import scenario.
In fact, you'll find a number of assumptions "baked in".
That *doesn't* mean that it's entirely useless.  It just means that you likely need to
do a bit more coding to get it to work for your application.

The first check in *doesn't* yet support the creation of the target file.  Presently, all
the output just goes to the console.